# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loop API",
    "version": "0.1.0",
    "description": "The Loop Autonomous Improvement Engine API. Collects signals, organizes work into issues, and dispatches AI agents with hydrated prompts.",
    "contact": {
      "name": "Loop by Dork Labs",
      "url": "https://www.looped.me"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5667",
      "description": "Local development"
    },
    {
      "url": "https://api.looped.me",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key passed as Bearer token in the Authorization header"
      }
    },
    "schemas": {
      "Issue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "number": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "[PostHog] error_rate: Login failures spiked"
          },
          "description": {
            "type": ["string", "null"],
            "example": "Auth failure rate exceeded 20% over the last hour."
          },
          "type": {
            "type": "string",
            "enum": ["signal", "hypothesis", "plan", "task", "monitor"],
            "example": "signal"
          },
          "status": {
            "type": "string",
            "enum": ["triage", "backlog", "todo", "in_progress", "done", "canceled"],
            "example": "triage"
          },
          "priority": {
            "type": "integer",
            "minimum": 0,
            "maximum": 4,
            "example": 2
          },
          "parentId": {
            "type": ["string", "null"],
            "example": null
          },
          "projectId": {
            "type": ["string", "null"],
            "example": null
          },
          "signalSource": {
            "type": ["string", "null"],
            "example": "posthog"
          },
          "signalPayload": {
            "type": ["object", "null"],
            "additionalProperties": {},
            "example": {
              "event": "error_rate",
              "value": 0.22
            }
          },
          "hypothesis": {
            "$ref": "#/components/schemas/HypothesisData"
          },
          "agentSessionId": {
            "type": ["string", "null"],
            "example": null
          },
          "agentSummary": {
            "type": ["string", "null"],
            "example": null
          },
          "commits": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/CommitRef"
            }
          },
          "pullRequests": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/PullRequestRef"
            }
          },
          "completedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": null
          }
        },
        "required": [
          "id",
          "number",
          "title",
          "description",
          "type",
          "status",
          "priority",
          "parentId",
          "projectId",
          "signalSource",
          "signalPayload",
          "hypothesis",
          "agentSessionId",
          "agentSummary",
          "commits",
          "pullRequests",
          "completedAt",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/issues": {
      "get": {
        "tags": ["Issues"],
        "summary": "List issues",
        "description": "Returns a paginated list of issues, filterable by status, type, projectId, labelId, priority, and parentId.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated list of statuses to filter by",
              "example": "todo,in_progress"
            },
            "required": false,
            "description": "Comma-separated list of statuses to filter by",
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated list of types to filter by",
              "example": "signal,task"
            },
            "required": false,
            "description": "Comma-separated list of types to filter by",
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "cuid2_proj"
            },
            "required": false,
            "name": "projectId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "cuid2_label"
            },
            "required": false,
            "name": "labelId",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "maximum": 4,
              "example": 2
            },
            "required": false,
            "name": "priority",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "cuid2_parent"
            },
            "required": false,
            "name": "parentId",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "example": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "example": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of issues",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Issue"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  },
                  "required": ["data", "total"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — invalid or missing Bearer token"
          }
        }
      },
      "post": {
        "tags": ["Issues"],
        "summary": "Create an issue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "example": "Investigate login failure spike"
                  },
                  "description": {
                    "type": "string",
                    "example": "Auth failure rate exceeded 20% over the last hour."
                  },
                  "type": {
                    "type": "string",
                    "enum": ["signal", "hypothesis", "plan", "task", "monitor"],
                    "example": "signal"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["triage", "backlog", "todo", "in_progress", "done", "canceled"],
                    "default": "triage",
                    "example": "triage"
                  },
                  "priority": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 4,
                    "default": 0,
                    "example": 2
                  },
                  "parentId": {
                    "type": "string"
                  },
                  "projectId": {
                    "type": "string"
                  },
                  "signalSource": {
                    "type": "string",
                    "example": "posthog"
                  },
                  "signalPayload": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "hypothesis": {
                    "$ref": "#/components/schemas/HypothesisData"
                  },
                  "labelIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": []
                  }
                },
                "required": ["title", "type"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Issue created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Issue"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation error or parent hierarchy violation"
          }
        }
      }
    },
...
    "/api/labels": {
      "get": {
        "tags": ["Labels"],
        "summary": "List labels",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "description": "Maximum records to return (1–200)",
              "example": 50
            },
            "required": false,
            "description": "Maximum records to return (1–200)",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "description": "Number of records to skip",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of labels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Label"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  },
                  "required": ["data", "total"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Labels"],
        "summary": "Create a label",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "frontend"
                  },
                  "color": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50,
                    "example": "#ff6b6b"
                  }
                },
                "required": ["name", "color"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Label created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Label"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "409": {
            "description": "Label name already exists"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List projects",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "description": "Maximum records to return (1–200)",
              "example": 50
            },
            "required": false,
            "description": "Maximum records to return (1–200)",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "description": "Number of records to skip",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  },
                  "required": ["data", "total"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a project",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "example": "Auth Reliability"
                  },
                  "description": {
                    "type": "string",
                    "example": "Improve authentication reliability."
                  },
                  "status": {
                    "type": "string",
                    "enum": ["backlog", "planned", "active", "paused", "completed", "canceled"],
                    "default": "backlog",
                    "example": "active"
                  },
                  "health": {
                    "type": "string",
                    "enum": ["on_track", "at_risk", "off_track"],
                    "default": "on_track",
                    "example": "on_track"
                  },
                  "goalId": {
                    "type": "string"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Project"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "422": {
            "description": "Validation error or goal not found"
          }
        }
      }
    },
...
    "/api/goals": {
      "get": {
        "tags": ["Goals"],
        "summary": "List goals",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "description": "Maximum records to return (1–200)",
              "example": 50
            },
            "required": false,
            "description": "Maximum records to return (1–200)",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "description": "Number of records to skip",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of goals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Goal"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  },
                  "required": ["data", "total"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Goals"],
        "summary": "Create a goal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "example": "Reduce login failure rate to < 5%"
                  },
                  "description": {
                    "type": "string"
                  },
                  "metric": {
                    "type": "string",
                    "example": "login_failure_rate"
                  },
                  "targetValue": {
                    "type": "number",
                    "example": 0.05
                  },
                  "currentValue": {
                    "type": "number",
                    "example": 0.18
                  },
                  "unit": {
                    "type": "string",
                    "example": "ratio"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["active", "achieved", "abandoned"],
                    "default": "active",
                    "example": "active"
                  },
                  "projectId": {
                    "type": "string"
                  }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Goal created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Goal"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/signals": {
      "post": {
        "tags": ["Signals"],
        "summary": "Ingest a signal",
        "description": "Creates a signal and atomically creates a linked triage issue.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "source": {
                    "type": "string",
                    "minLength": 1,
                    "example": "posthog"
                  },
                  "sourceId": {
                    "type": "string",
                    "example": "evt_12345"
                  },
                  "type": {
                    "type": "string",
                    "minLength": 1,
                    "example": "error_rate"
                  },
                  "severity": {
                    "type": "string",
                    "enum": ["low", "medium", "high", "critical"],
                    "example": "high"
                  },
                  "payload": {
                    "type": "object",
                    "additionalProperties": {},
                    "example": {
                      "event": "error_rate",
                      "value": 0.22
                    }
                  },
                  "projectId": {
                    "type": "string"
                  }
                },
                "required": ["source", "type", "severity", "payload"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signal and linked issue created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "signal": {
                          "$ref": "#/components/schemas/Signal"
                        },
                        "issue": {
                          "$ref": "#/components/schemas/Issue"
                        }
                      },
                      "required": ["signal", "issue"]
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/templates": {
      "get": {
        "tags": ["Templates"],
        "summary": "List prompt templates",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "description": "Maximum records to return (1–200)",
              "example": 50
            },
            "required": false,
            "description": "Maximum records to return (1–200)",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "description": "Number of records to skip",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of prompt templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PromptTemplate"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  },
                  "required": ["data", "total"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Templates"],
        "summary": "Create a prompt template",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "example": "signal-triage"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "example": "Signal Triage"
                  },
                  "description": {
                    "type": "string",
                    "example": "Default template for triaging signals."
                  },
                  "conditions": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": ["signal", "hypothesis", "plan", "task", "monitor"]
                      },
                      "signalSource": {
                        "type": "string"
                      },
                      "labels": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "projectId": {
                        "type": "string"
                      },
                      "hasFailedSessions": {
                        "type": "boolean"
                      },
                      "hypothesisConfidence": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 1
                      }
                    },
                    "default": {},
                    "description": "Conditions that determine when this template is selected",
                    "example": {
                      "type": "signal"
                    }
                  },
                  "specificity": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100,
                    "default": 10,
                    "example": 10
                  },
                  "projectId": {
                    "type": "string"
                  }
                },
                "required": ["slug", "name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PromptTemplate"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "409": {
            "description": "Template slug already exists"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/prompt-reviews": {
      "post": {
        "tags": ["Prompt Reviews"],
        "summary": "Submit a prompt review",
        "description": "Submits a quality review for a prompt version. Updates the version's EWMA review score and triggers an improvement issue if quality degrades below threshold.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "versionId": {
                    "type": "string",
                    "example": "cuid2_abc123",
                    "description": "The prompt version being reviewed"
                  },
                  "issueId": {
                    "type": "string",
                    "example": "cuid2_abc123",
                    "description": "The issue this prompt was used for"
                  },
                  "clarity": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 4
                  },
                  "completeness": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 5
                  },
                  "relevance": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 3
                  },
                  "feedback": {
                    "type": "string",
                    "example": "Instructions were clear but missing error handling steps."
                  },
                  "authorType": {
                    "type": "string",
                    "enum": ["human", "agent"],
                    "example": "agent"
                  }
                },
                "required": [
                  "versionId",
                  "issueId",
                  "clarity",
                  "completeness",
                  "relevance",
                  "authorType"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PromptReview"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "404": {
            "description": "Version not found"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
...
  },
...
}